souoc

+++ so uoc.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
 int main(){
-	long long n;
+	int n;
 	scanf("%d",&n);
-	long long i;
-	long long so_uoc = 0;
+	int i;
+	int so_uoc = 0;
 	for(i = 1; i <= n; i++){
 		if(n % i == 0){
 			so_uoc ++;
 		}
 	}
-	printf("%lld",so_uoc);
+	printf("%d",so_uoc);
 	return 0;
 }
